+++ internal/commands/all/list_test.go
@@ -26,6 +26,7 @@ func mockListResponses(mService *smock.Service) {
 	mService.On("GetKubernetesClusters", mock.Anything).Return(nil, nil)
 	mService.On("GetLoadBalancers", mock.Anything).Return(nil, nil)
 	mService.On("GetFileStorages", mock.Anything).Return(nil, nil)
+	mService.On("GetGateways", mock.Anything).Return(nil, nil)
 }
 
 var networks = &upcloud.Networks{Networks: []upcloud.Network{